+++ R/Lib_MapBetaDiversity.R
@@ -3,8 +3,8 @@
 # Lib_MapBetaDiversity.R
 # ==============================================================================
 # PROGRAMMERS:
-# Jean-Baptiste FERET <jb.feret@irstea.fr>
-# Copyright 2018/07 Jean-Baptiste FERET
+# Jean-Baptiste FERET <jb.feret@teledetection.fr>
+# Copyright 2020/06 Jean-Baptiste FERET
 # ==============================================================================
 # This Library computes bray curtis dissimilarity among spatial units based on
 # spectral species distribution file and generates a RGB representation with NMDS